feat: implementa logica de deletar autor no service, nao permitindo excluir caso houver livros cadastrados

+++ src/main/java/io/github/luizeduardotsdev/livrariaapi/service/AutorService.java
@@ -1,7 +1,9 @@
 package io.github.luizeduardotsdev.livrariaapi.service;
 
+import io.github.luizeduardotsdev.livrariaapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.luizeduardotsdev.livrariaapi.model.Autor;
 import io.github.luizeduardotsdev.livrariaapi.repository.AutorRepository;
+import io.github.luizeduardotsdev.livrariaapi.repository.LivroRepository;
 import io.github.luizeduardotsdev.livrariaapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
@@ -14,10 +16,12 @@ public class AutorService {
 
     private final AutorRepository autorRepository;
     private final AutorValidator autorValidator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository autorRepository, AutorValidator autorValidator) {
+    public AutorService(AutorRepository autorRepository, AutorValidator autorValidator, LivroRepository livroRepository) {
         this.autorRepository = autorRepository;
         this.autorValidator = autorValidator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor) {
@@ -38,6 +42,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("nao é permitido exlcuir um autor que posuui livros cadastrados");
+        }
         autorRepository.delete(autor);
     }
 
@@ -55,6 +62,10 @@ public class AutorService {
         }
 
         return autorRepository.findAll();
-
     }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
+    }
+
 }
